Use angular not linear coordinates in geostationary
Fixes #230

The initial definition for this projection was agreed on in May 2012 but
was not adopted in the CF document until v1.7 was released in September
2017. At that time, new generations of satellite missions were being
launched that generate data using this now deprecated method, including
GOES-R (NOAA, operational in December 2017). Subsequent missions, e.g.
Meteosat Third Generation (EUMETSAT) are encouraged to use CF-1.8
onwards.